Rename NameSpaces and Folders from "Newss" to "News"
