+++ Streetcode/Streetcode.BLL/MediatR/News/Delete/DeleteNewsCommand.cs
@@ -1,0 +1,7 @@
+﻿using FluentResults;
+using MediatR;
+
+namespace Streetcode.BLL.MediatR.News.Delete
+{
+    public record DeleteNewsCommand(int id) : IRequest<Result<Unit>>;
+}
