+++ Streetcode/Streetcode.BLL/MediatR/News/SortedByDateTime/SortedByDateTimeHandler.cs
@@ -1,0 +1,53 @@
+﻿using AutoMapper;
+using FluentResults;
+using MediatR;
+using Streetcode.BLL.DTO.News;
+using Streetcode.BLL.Interfaces.BlobStorage;
+using Microsoft.EntityFrameworkCore;
+using Streetcode.DAL.Repositories.Interfaces.Base;
+using Streetcode.BLL.Interfaces.Logging;
+using Streetcode.BLL.DTO.AdditionalContent.Subtitles;
+
+namespace Streetcode.BLL.MediatR.News.SortedByDateTime
+{
+    public class SortedByDateTimeHandler : IRequestHandler<SortedByDateTimeQuery, Result<List<NewsDTO>>>
+    {
+        private readonly IRepositoryWrapper _repositoryWrapper;
+        private readonly IMapper _mapper;
+        private readonly IBlobService _blobService;
+        private readonly ILoggerService _logger;
+
+        public SortedByDateTimeHandler(IRepositoryWrapper repositoryWrapper, IMapper mapper, IBlobService blobService, ILoggerService logger)
+        {
+            _repositoryWrapper = repositoryWrapper;
+            _mapper = mapper;
+            _blobService = blobService;
+            _logger = logger;
+        }
+
+        public async Task<Result<List<NewsDTO>>> Handle(SortedByDateTimeQuery request, CancellationToken cancellationToken)
+        {
+            var news = await _repositoryWrapper.NewsRepository.GetAllAsync(
+                include: cat => cat.Include(img => img.Image));
+            if (news == null)
+            {
+                const string errorMsg = "There are no news in the database";
+                _logger.LogError(request, errorMsg);
+                return Result.Fail(errorMsg);
+            }
+
+            var newsDTOs = _mapper.Map<IEnumerable<NewsDTO>>(news).OrderByDescending(x => x.CreationDate).ToList();
+
+            newsDTOs = newsDTOs
+                .Where(dto => dto.Image is not null)
+                .Select(dto =>
+                {
+                    dto.Image!.Base64 = _blobService.FindFileInStorageAsBase64(dto.Image.BlobName!);
+                    return dto;
+                })
+                .ToList();
+
+            return Result.Ok(newsDTOs);
+        }
+    }
+}
