--- Streetcode/Streetcode.BLL/MediatR/Newss/Create/CreateNewsCommand.cs
@@ -1,8 +1,0 @@
-﻿using FluentResults;
-using MediatR;
-using Streetcode.BLL.DTO.News;
-
-namespace Streetcode.BLL.MediatR.Newss.Create
-{
-    public record CreateNewsCommand(NewsDTO newNews) : IRequest<Result<NewsDTO>>;
-}
