--- Streetcode/Streetcode.BLL/MediatR/Newss/Delete/DeleteNewsCommand.cs
@@ -1,7 +1,0 @@
-﻿using FluentResults;
-using MediatR;
-
-namespace Streetcode.BLL.MediatR.Newss.Delete
-{
-    public record DeleteNewsCommand(int id) : IRequest<Result<Unit>>;
-}
