--- Streetcode/Streetcode.BLL/MediatR/Newss/GetByUrl/GetNewsByUrlQuery.cs
@@ -1,8 +1,0 @@
-﻿using FluentResults;
-using MediatR;
-using Streetcode.BLL.DTO.News;
-
-namespace Streetcode.BLL.MediatR.Newss.GetByUrl
-{
-    public record GetNewsByUrlQuery(string url) : IRequest<Result<NewsDTO>>;
-}
