--- Streetcode/Streetcode.BLL/MediatR/Newss/GetNewsAndLinksByUrl/GetNewsAndLinksByUrlQuery.cs
@@ -1,8 +1,0 @@
-﻿using FluentResults;
-using MediatR;
-using Streetcode.BLL.DTO.News;
-
-namespace Streetcode.BLL.MediatR.Newss.GetNewsAndLinksByUrl
-{
-    public record GetNewsAndLinksByUrlQuery(string url) : IRequest<Result<NewsDTOWithURLs>>;
-}
