--- Streetcode/Streetcode.BLL/MediatR/Newss/SortedByDateTime/SortedByDateTimeQuery.cs
@@ -1,8 +1,0 @@
-﻿using FluentResults;
-using MediatR;
-using Streetcode.BLL.DTO.News;
-
-namespace Streetcode.BLL.MediatR.Newss.SortedByDateTime
-{
-    public record SortedByDateTimeQuery() : IRequest<Result<List<NewsDTO>>>;
-}
